Enable Vulkan for WGPU runner
Metal doesn't support SPIR-V shader passthrough, so on macOS using WGPU runner you get following error with any example:

```
Features SPIRV_SHADER_PASSTHROUGH are required but not enabled on the device
```

Vulkan supports shader passthrough, but requires compatibility layer of MoltenVK to work, which needs to be installed on the host machine. Currently this feature is only available on master.