Switch `rustc_codgen_spirv` to static lib
Was set as a `dylib` but not really sure why. Having it as a static library enables dependencies, like `spriv-builder`, to start building directly after the frontend section of the build of the crate is done. Rather than having to wait for the codegen as well.

Saw clean ´spirv-builder` build times go from 46.3 -> 45.8 seconds with this. Have seen bigger benefits when combined with other optimizations as well.